Split Authenticator logic from SSOAuthenticator class
